+++ src/audio/eq_fir/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_local_sources(sof eq_fir.c fir_hifi2ep.c fir_hifi3.c fir.c)
