+++ src/audio/eq_iir/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_local_sources(sof eq_iir.c iir.c)
