+++ src/audio/selector/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_local_sources(sof selector_generic.c selector.c)
