+++ src/audio/src/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_local_sources(sof src_generic.c src_hifi2ep.c src_hifi3.c src.c)
