+++ src/audio/src/src_generic.c
@@ -36,10 +36,9 @@
 #include <stdint.h>
 #include <sof/alloc.h>
 #include <sof/audio/format.h>
+#include <sof/audio/src/src_config.h>
+#include <sof/audio/src/src.h>
 #include <sof/math/numbers.h>
-
-#include "src_config.h"
-#include "src.h"
 
 #if SRC_GENERIC
