+++ src/audio/src/src_hifi3.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, Intel Corporation
+ * Copyright (c) 2016, Intel Corporation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,195 +29,201 @@
  *
  */
 
-/* HiFi EP optimized code parts for SRC */
+/* HiFi3 optimized code parts for SRC */
 
 #include <stdint.h>
 #include <sof/alloc.h>
 #include <sof/audio/format.h>
+#include <sof/audio/src/src_config.h>
+#include <sof/audio/src/src.h>
 #include <sof/math/numbers.h>
 
-#include "src_config.h"
-#include "src.h"
-
-#if SRC_HIFIEP
+#if SRC_HIFI3
 
 #include <xtensa/config/defs.h>
-#include <xtensa/tie/xt_hifi2.h>
+#include <xtensa/tie/xt_hifi3.h>
 
-/* HiFi EP has
- * 4x 56 bit registers in register file Q
- * 8x 48 bit registers in register file P
+/* HiFi3 has
+ * 16x 64 bit registers in register file AE_DR
  */
 
 #if SRC_SHORT /* 16 bit coefficients version */
 
-static inline void fir_filter(ae_q32s *rp, const void *cp, ae_q32s *wp0,
+static inline void fir_filter(ae_f32 *rp, const void *cp, ae_f32 *wp0,
 			      const int taps_div_4, const int shift,
 			      const int nch)
 {
 	/* This function uses
-	 * 2x 56 bit registers Q,
-	 * 4x 48 bit registers P
+	 * 6x 64 bit registers
 	 * 3x integers
-	 * 4x address pointers,
+	 * 5x address pointers,
 	 */
-	ae_q56s a0;
-	ae_q56s a1;
-	ae_p24x2f data2;
-	ae_p24x2f coef2;
-	ae_p24x2f p0;
-	ae_p24x2f p1;
-	ae_p16x2s *coefp;
-	ae_p24x2f *dp = (ae_p24x2f *)rp;
-	ae_p24x2f *dp0;
-	ae_q32s *wp = wp0;
+	ae_f64 a0;
+	ae_f64 a1;
+	ae_valign u;
+	ae_f16x4 coef4;
+	ae_f32x2 d0;
+	ae_f32x2 d1;
+	ae_f32x2 data2;
+	ae_f16x4 *coefp;
+	ae_f32x2 *dp;
+	ae_f32 *dp0;
+	ae_f32 *dp1;
 	int i;
 	int j;
-	const int inc = sizeof(ae_p24x2f);
+	ae_f32 *wp = wp0;
+	const int inc = nch * sizeof(int32_t);
 
-	/* 2ch FIR case */
 	if (nch == 2) {
 		/* Move data pointer back by one sample to start from right
 		 * channel sample. Discard read value p0.
 		 */
-		AE_LP24F_C(p0, dp, -sizeof(ae_p24f));
+		dp = (ae_f32x2 *)rp;
+		AE_L32_XC(d0, (ae_f32 *)dp, -sizeof(ae_f32));
 
 		/* Reset coefficient pointer and clear accumulator */
-		coefp = (ae_p16x2s *)cp;
-		a0 = AE_ZEROQ56();
-		a1 = AE_ZEROQ56();
+		coefp = (ae_f16x4 *)cp;
+		a0 = AE_ZERO64();
+		a1 = AE_ZERO64();
 
 		/* Compute FIR filter for current channel with four
-		 * taps per every loop iteration.  Two coefficients
+		 * taps per every loop iteration.  Four coefficients
 		 * are loaded simultaneously. Data is read
 		 * from interleaved buffer with stride of channels
 		 * count.
 		 */
 		for (i = 0; i < taps_div_4; i++) {
-			/* Load two coefficients. Coef2_h contains tap *coefp
-			 * and coef2_l contains the next tap.
-			 */
-			coef2 = AE_LP16X2F_I(coefp, 0);
-			coefp++;
+			/* Load four coefficients */
+			AE_LA16X4_IP(coef4, u, coefp);
 
 			/* Load two data samples from two channels */
-			AE_LP24X2F_C(p0, dp, inc); /* r0, l0 */
-			AE_LP24X2F_C(p1, dp, inc); /* r1, l1 */
+			AE_L32X2_XC(d0, dp, inc); /* r0, l0 */
+			AE_L32X2_XC(d1, dp, inc); /* r1, l1 */
 
-			/* Select to d0 successive left channel samples, to d1
-			 * successive right channel samples. Then accumulate
-			 * data2_h * coef2_h + data2_l * coef2_l. The Q1.31
-			 * data and Q1.15 coefficients are used as 24 bits as
-			 * Q1.23 values.
+			/* Select to data2 sequential samples from a channel
+			 * and then accumulate to a0 and a1
+			 * data2_h * coef4_3 + data2_l * coef4_2.
+			 * The data is 32 bits Q1.31 and coefficient 16 bits
+			 * Q1.15. The accumulators are Q17.47.
 			 */
-			data2 = AE_SELP24_LL(p0, p1);
-			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
-			data2 = AE_SELP24_HH(p0, p1);
-			AE_MULAAFP24S_HH_LL(a1, data2, coef2);
+			data2 = AE_SEL32_LL(d0, d1); /* l0, l1 */
+			AE_MULAAFD32X16_H3_L2(a0, data2, coef4);
+			data2 = AE_SEL32_HH(d0, d1); /* r0, r1 */
+			AE_MULAAFD32X16_H3_L2(a1, data2, coef4);
 
-			/* Repeat for next two taps */
-			coef2 = AE_LP16X2F_I(coefp, 0);
-			coefp++;
-			AE_LP24X2F_C(p0, dp, inc); /* r2, l2 */
-			AE_LP24X2F_C(p1, dp, inc); /* r3, l3 */
-			data2 = AE_SELP24_LL(p0, p1);
-			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
-			data2 = AE_SELP24_HH(p0, p1);
-			AE_MULAAFP24S_HH_LL(a1, data2, coef2);
+			/* Load two data samples from two channels */
+			AE_L32X2_XC(d0, dp, inc); /* r2, l2 */
+			AE_L32X2_XC(d1, dp, inc); /* r3, l3 */
+
+			/* Accumulate
+			 * data2_h * coef4_1 + data2_l * coef4_0.
+			 */
+			data2 = AE_SEL32_LL(d0, d1); /* l2, l3 */
+			AE_MULAAFD32X16_H1_L0(a0, data2, coef4);
+			data2 = AE_SEL32_HH(d0, d1); /* r2, r3 */
+			AE_MULAAFD32X16_H1_L0(a1, data2, coef4);
 		}
 
 		/* Scale FIR output with right shifts, round/saturate
 		 * to Q1.31, and store 32 bit output.
 		 */
-		AE_SQ32F_I(AE_ROUNDSQ32SYM(AE_SRAAQ56(a0, shift)), wp, 0);
-		AE_SQ32F_I(AE_ROUNDSQ32SYM(AE_SRAAQ56(a1, shift)), wp,
-			   sizeof(int32_t));
+		AE_S32_L_XP(AE_ROUND32F48SSYM(AE_SRAA64(a0, shift)), wp,
+			    sizeof(int32_t));
+		AE_S32_L_XP(AE_ROUND32F48SSYM(AE_SRAA64(a1, shift)), wp,
+			    sizeof(int32_t));
+
 		return;
 	}
 
+	dp1 = (ae_f32 *)rp;
 	for (j = 0; j < nch; j++) {
 		/* Copy pointer and advance to next ch with dummy load */
-		dp0 = dp;
-		AE_LP24F_C(p0, dp, -sizeof(ae_p24f));
+		dp0 = dp1;
+		AE_L32_XC(d0, dp1, -sizeof(ae_f32));
 
 		/* Reset coefficient pointer and clear accumulator */
-		coefp = (ae_p16x2s *)cp;
-		a0 = AE_ZEROQ56();
+		coefp = (ae_f16x4 *)cp;
+		a0 = AE_ZERO64();
 
 		/* Compute FIR filter for current channel with four
-		 * taps per every loop iteration.  Two coefficients
-		 * are loaded simultaneously. Data is read
-		 * from interleaved buffer with stride of channels
-		 * count.
+		 * taps per every loop iteration. Data is read from
+		 * interleaved buffer with stride of channels count.
 		 */
 		for (i = 0; i < taps_div_4; i++) {
-			/* Load two coefficients */
-			coef2 = *coefp++;
+			/* Load four coefficients */
+			AE_LA16X4_IP(coef4, u, coefp);
 
-			/* Load two data samples */
-			AE_LP24F_C(p0, dp0, inc);
-			AE_LP24F_C(p1, dp0, inc);
+			/* Load two data samples, place to high and
+			 * low of data2.
+			 */
+			AE_L32_XC(d0, dp0, inc);
+			AE_L32_XC(d1, dp0, inc);
+			data2 = AE_SEL32_LL(d0, d1);
 
-			/* Pack p0 and p1 to data2_h and data2_l */
-			data2 = AE_SELP24_LL(p0, p1);
+			/* Accumulate
+			 * data2_h * coef4_3 + data2_l* coef4_2.
+			 * The data is 32 bits Q1.31 and coefficient 16 bits
+			 * Q1.15. The accumulator is Q17.47.
+			 */
+			AE_MULAAFD32X16_H3_L2(a0, data2, coef4);
 
-			/* Accumulate data2_h * coef2_h + data2_l * coef2_l */
-			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
+			/* Repeat with next two samples */
+			AE_L32_XC(d0, dp0, inc);
+			AE_L32_XC(d1, dp0, inc);
+			data2 = AE_SEL32_LL(d0, d1);
 
-			/* Repeat for next two filter taps */
-			coef2 = *coefp++;
-			AE_LP24F_C(p0, dp0, inc);
-			AE_LP24F_C(p1, dp0, inc);
-			data2 = AE_SELP24_LL(p0, p1);
-			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
+			/* Accumulate
+			 * data2_h * coef4_1 + data2_l * coef4_0.
+			 */
+			AE_MULAAFD32X16_H1_L0(a0, data2, coef4);
 		}
 
-		/* Scale FIR output with right shifts, round/saturate
+		/* Scale FIR output with right shifts, round/saturate Q17.47
 		 * to Q1.31, and store 32 bit output. Advance write
 		 * pointer to next sample.
 		 */
-		AE_SQ32F_I(AE_ROUNDSQ32SYM(AE_SRAAQ56(a0, shift)), wp, 0);
-		wp++;
+		AE_S32_L_XP(AE_ROUND32F48SSYM(AE_SRAA64(a0, shift)), wp,
+			    sizeof(int32_t));
 	}
 }
 
 #else /* 32bit coefficients version */
 
-static inline void fir_filter(ae_q32s *rp, const void *cp, ae_q32s *wp0,
+static inline void fir_filter(ae_f32 *rp, const void *cp, ae_f32 *wp0,
 			      const int taps_div_4, const int shift,
 			      const int nch)
 {
 	/* This function uses
-	 * 2x 56 bit registers Q,
-	 * 4x 48 bit registers P
+	 * 6x 64 bit registers
 	 * 3x integers
-	 * 4x address pointers,
+	 * 5x address pointers,
 	 */
-	ae_q56s a0;
-	ae_q56s a1;
-	ae_p24x2f p0;
-	ae_p24x2f p1;
-	ae_p24x2f data2;
-	ae_p24x2f coef2;
-	ae_p24x2f *coefp;
-	ae_p24x2f *dp = (ae_p24x2f *)rp;
-	ae_p24x2f *dp0;
-	ae_q32s *wp = wp0;
+	ae_f64 a0;
+	ae_f64 a1;
+	ae_f24x2 data2 = AE_ZERO24();
+	ae_f24x2 coef2 = AE_ZERO24();
+	ae_f24x2 d0 = AE_ZERO24();
+	ae_f24x2 d1 = AE_ZERO24();
+	ae_f24x2 *coefp;
+	ae_f24x2 *dp;
+	ae_f24 *dp1;
+	ae_f24 *dp0;
 	int i;
 	int j;
-	const int inc = sizeof(ae_p24x2f);
+	ae_f32 *wp = wp0;
+	const int inc = nch * sizeof(int32_t);
 
-	/* 2ch FIR case */
 	if (nch == 2) {
 		/* Move data pointer back by one sample to start from right
 		 * channel sample. Discard read value p0.
 		 */
-		AE_LP24F_C(p0, dp, -sizeof(ae_p24f));
+		dp = (ae_f24x2 *)rp;
+		AE_L32F24_XC(d0, (ae_f24 *)dp, -sizeof(ae_f24));
 
 		/* Reset coefficient pointer and clear accumulator */
-		coefp = (ae_p24x2f *)cp;
-		a0 = AE_ZEROQ56();
-		a1 = AE_ZEROQ56();
+		coefp = (ae_f24x2 *)cp;
+		a0 = AE_ZERO64();
+		a1 = AE_ZERO64();
 
 		/* Compute FIR filter for current channel with four
 		 * taps per every loop iteration.  Two coefficients
@@ -230,109 +236,106 @@ static inline void fir_filter(ae_q32s *rp, const void *cp, ae_q32s *wp0,
 			 * and coef2_l contains the next tap.
 			 */
 			/* TODO: Ensure coefficients are 64 bits aligned */
-			coef2 = AE_LP24X2F_I(coefp, 0);
-			coefp++;
+			AE_L32X2F24_IP(coef2, coefp, sizeof(ae_f24x2));
 
 			/* Load two data samples from two channels */
-			AE_LP24X2F_C(p0, dp, inc); /* r0, l0 */
-			AE_LP24X2F_C(p1, dp, inc); /* r1, l1 */
+			AE_L32X2F24_XC(d0, dp, inc); /* r0, l0 */
+			AE_L32X2F24_XC(d1, dp, inc); /* r1, l1 */
 
 			/* Select to d0 successive left channel samples, to d1
-			 * successive right channel samples.
-			 */
-
-			/* Accumulate to m
+			 * successive right channel samples. Then Accumulate
+			 * to a0 and a1
 			 * data2_h * coef2_h + data2_l * coef2_l. The Q1.31
 			 * data and Q1.15 coefficients are used as 24 bits as
 			 * Q1.23 values.
 			 */
-			data2 = AE_SELP24_LL(p0, p1);
+			data2 = AE_SELP24_LL(d0, d1);
 			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
-			data2 = AE_SELP24_HH(p0, p1);
+			data2 = AE_SELP24_HH(d0, d1);
 			AE_MULAAFP24S_HH_LL(a1, data2, coef2);
 
 			/* Repeat for next two taps */
-			coef2 = AE_LP24X2F_I(coefp, 0);
-			coefp++;
-			AE_LP24X2F_C(p0, dp, inc); /* r2, l2 */
-			AE_LP24X2F_C(p1, dp, inc); /* r3, l3 */
-			data2 = AE_SELP24_LL(p0, p1);
+			AE_L32X2F24_IP(coef2, coefp, sizeof(ae_f24x2));
+			AE_L32X2F24_XC(d0, dp, inc); /* r2, l2 */
+			AE_L32X2F24_XC(d1, dp, inc); /* r3, l3 */
+			data2 = AE_SELP24_LL(d0, d1);
 			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
-			data2 = AE_SELP24_HH(p0, p1);
+			data2 = AE_SELP24_HH(d0, d1);
 			AE_MULAAFP24S_HH_LL(a1, data2, coef2);
 		}
 
 		/* Scale FIR output with right shifts, round/saturate
 		 * to Q1.31, and store 32 bit output.
 		 */
-		AE_SQ32F_I(AE_ROUNDSQ32SYM(AE_SRAAQ56(a0, shift)), wp, 0);
-		AE_SQ32F_I(AE_ROUNDSQ32SYM(AE_SRAAQ56(a1, shift)), wp,
-			   sizeof(int32_t));
+		AE_S32_L_XP(AE_ROUND32F48SSYM(AE_SRAA64(a0, shift)), wp,
+			    sizeof(int32_t));
+		AE_S32_L_XP(AE_ROUND32F48SSYM(AE_SRAA64(a1, shift)), wp,
+			    sizeof(int32_t));
+
 		return;
 	}
 
+	dp1 = (ae_f24 *)rp;
 	for (j = 0; j < nch; j++) {
 		/* Copy pointer and advance to next ch with dummy load */
-		dp0 = dp;
-		AE_LP24F_C(p0, dp, -sizeof(ae_p24f));
+		dp0 = dp1;
+		AE_L32F24_XC(data2, dp1, -sizeof(ae_f24));
 
 		/* Reset coefficient pointer and clear accumulator */
-		coefp = (ae_p24x2f *)cp;
-		a0 = AE_ZEROQ56();
+		coefp = (ae_f24x2 *)cp;
+		a0 = AE_ZERO64();
 
 		/* Compute FIR filter for current channel with four
-		 * taps per every loop iteration.  Two coefficients
-		 * are loaded simultaneously. Data is read
-		 * from interleaved buffer with stride of channels
-		 * count.
+		 * taps per every loop iteration. Data is read from
+		 * interleaved buffer with stride of channels count.
 		 */
 		for (i = 0; i < taps_div_4; i++) {
 			/* Load two coefficients */
 			coef2 = *coefp++;
 
-			/* Load two data samples and place them to L and H of
-			 * data2.
+			/* Load two data samples, place to high and
+			 * low of data2.
 			 */
-			AE_LP24F_C(p0, dp0, inc);
-			AE_LP24F_C(p1, dp0, inc);
-			data2 = AE_SELP24_LH(p0, p1);
+			AE_L32F24_XC(d0, dp0, inc);
+			AE_L32F24_XC(d1, dp0, inc);
+			data2 = AE_SELP24_LL(d0, d1);
 
-			/* Accumulate to m
-			 * data2_h * coef2_h + data2_l * coef2_l. The Q1.31
-			 * data and coefficients are used as the most
-			 * significant 24 bits as Q1.23 values.
+			/* Accumulate to data2_h * coef2_h +
+			 * data2_l*coef2_l. The Q1.31 bit data is used
+			 * as Q1.23 from MSB side bits of the 32 bit
+			 * word. The accumulator m is Q17.47.
 			 */
-			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
+			AE_MULAAFD24_HH_LL(a0, data2, coef2);
 
-			/* Repeat for next two filter taps */
+			/* Repeat the same for next two filter taps */
 			coef2 = *coefp++;
-			AE_LP24F_C(p0, dp0, inc);
-			AE_LP24F_C(p1, dp0, inc);
-			data2 = AE_SELP24_LH(p0, p1);
-			AE_MULAAFP24S_HH_LL(a0, data2, coef2);
+			AE_L32F24_XC(d0, dp0, inc);
+			AE_L32F24_XC(d1, dp0, inc);
+			data2 = AE_SELP24_LL(d0, d1);
+			AE_MULAAFD24_HH_LL(a0, data2, coef2);
 		}
 
-		/* Scale FIR output with right shifts, round/saturate
+		/* Scale FIR output with right shifts, round/saturate Q17.47
 		 * to Q1.31, and store 32 bit output. Advance write
 		 * pointer to next sample.
 		 */
-		AE_SQ32F_I(AE_ROUNDSQ32SYM(AE_SRAAQ56(a0, shift)), wp, 0);
-		wp++;
+		AE_S32_L_XP(AE_ROUND32F48SSYM(AE_SRAA64(a0, shift)), wp,
+			    sizeof(int32_t));
 	}
 }
+
 #endif /* 32bit coefficients version */
 
 void src_polyphase_stage_cir(struct src_stage_prm *s)
 {
 	/* This function uses
-	 *  1x 56 bit registers Q,
-	 *  0x 48 bit registers P,
+	 *  1x 64 bit registers
 	 * 16x integers
 	 * 11x address pointers,
 	 */
-	ae_q56s q;
-	ae_q32s *rp;
-	ae_q32s *wp;
+	ae_int32x2 q = AE_ZERO32();
+	ae_f32 *rp;
+	ae_f32 *wp;
 	int i;
 	int n;
 	int m;
@@ -379,14 +382,14 @@ void src_polyphase_stage_cir(struct src_stage_prm *s)
 			n_min = (m < n_wrap_buf) ? m : n_wrap_buf;
 			m -= n_min;
 			for (i = 0; i < n_min; i++) {
-				/* Load 32 bits sample to accumulator */
-				q = AE_LQ32F_I((ae_q32s *)x_rptr++, 0);
-
-				/* Saturating shift left */
-				q = AE_SLLASQ56S(q, s->shift);
+				/* Load 32 bits sample to accumulator,
+				 * advance pointer, shift left with saturate.
+				 */
+				AE_L32_XP(q, (ae_int32 *)x_rptr, sz);
+				q = AE_SLAA32(q, s->shift);
 
 				/* Store to circular buffer, advance pointer */
-				AE_SQ32F_C(q, (ae_q32s *)fir->fir_wp, n_sz);
+				AE_S32_L_XC(q, (ae_int32 *)fir->fir_wp, n_sz);
 			}
 
 			/* Check for wrap */
@@ -395,17 +398,17 @@ void src_polyphase_stage_cir(struct src_stage_prm *s)
 
 		/* Do filter */
 		cp = cfg->coefs; /* Reset to 1st coefficient */
-		rp = (ae_q32s *)fir->fir_wp;
+		rp = (ae_f32 *)fir->fir_wp;
 
 		/* Do circular modification to pointer rp by amount of
 		 * rewind to to data start. Loaded value q is discarded.
 		 */
-		AE_LQ32F_C(q, (ae_q32s *)rp, rewind_sz);
+		AE_L32_XC(q, rp, rewind_sz);
 
 		/* Reset FIR write pointer and compute all polyphase
 		 * sub-filters.
 		 */
-		wp = (ae_q32s *)fir->out_rp;
+		wp = (ae_f32 *)fir->out_rp;
 		for (i = 0; i < cfg->num_of_subfilters; i++) {
 			fir_filter(rp, cp, wp, taps_div_4, cfg->shift, nch);
 			wp += nch_x_odm;
@@ -416,7 +419,7 @@ void src_polyphase_stage_cir(struct src_stage_prm *s)
 			 * channels x input delay multiplier. Loaded value q
 			 * is discarded.
 			 */
-			AE_LQ32F_C(q, rp, nch_x_idm_sz);
+			AE_L32_XC(q, rp, nch_x_idm_sz);
 		}
 
 		/* Output */
@@ -430,14 +433,14 @@ void src_polyphase_stage_cir(struct src_stage_prm *s)
 			n_min = (m < n_wrap_buf) ? m : n_wrap_buf;
 			m -= n_min;
 			for (i = 0; i < n_min; i++) {
-				/* Circular load followed by shift right
-				 * for optional s24 and linear store
+				/* Circular load, shift right, linear store,
+				 * and advance read and write pointers.
 				 */
-				AE_LQ32F_C(q, (ae_q32s *)fir->out_rp, sz);
-				q = AE_SRAAQ56(q, s->shift);
-				AE_SQ32F_I(q, (ae_q32s *)y_wptr, 0);
-				y_wptr++;
+				AE_L32_XC(q, (ae_int32 *)fir->out_rp, sz);
+				q = AE_SRAA32(q, s->shift);
+				AE_S32_L_XP(q, (ae_int32 *)y_wptr, sz);
 			}
+
 			/* Check wrap */
 			src_inc_wrap(&y_wptr, y_end_addr, s->y_size);
 		}
@@ -449,19 +452,20 @@ void src_polyphase_stage_cir(struct src_stage_prm *s)
 void src_polyphase_stage_cir_s16(struct src_stage_prm *s)
 {
 	/* This function uses
-	 *  0x 56 bit registers Q,
-	 *  1x 48 bit registers P,
+	 *  2x 64 bit registers
 	 * 16x integers
 	 * 11x address pointers,
 	 */
-	ae_p24x2s d;
-	ae_q32s *rp;
-	ae_q32s *wp;
+	ae_int32x2 q = AE_ZERO32();
+	ae_int16x4 d = AE_ZERO16();
+	ae_f32 *rp;
+	ae_f32 *wp;
 	int i;
 	int n;
 	int m;
 	int n_wrap_buf;
 	int n_min;
+
 	struct src_state *fir = s->state;
 	struct src_stage *cfg = s->stage;
 	int32_t *fir_end = &fir->fir_delay[fir->fir_delay_size];
@@ -490,7 +494,7 @@ void src_polyphase_stage_cir_s16(struct src_stage_prm *s)
 #endif
 
 	for (n = 0; n < s->times; n++) {
-		/* Input data to filter */
+		/* Input data */
 		m = blk_in_words;
 
 		/* Setup circular buffer for FIR input data delay */
@@ -503,16 +507,14 @@ void src_polyphase_stage_cir_s16(struct src_stage_prm *s)
 			n_min = (m < n_wrap_buf) ? m : n_wrap_buf;
 			m -= n_min;
 			for (i = 0; i < n_min; i++) {
-				/* Load 16 bits sample to 24 bit register
-				 * and advance read pointer.
+				/* Load a 16 bits sample into d and left shift
+				 * by 16 into q, advance read and write
+				 * pointers.
 				 */
-				d = AE_LP16F_I((ae_p16s *)x_rptr, 0);
-				x_rptr++;
-
-				/* Store to 32 bit circular buffer, advance
-				 * write pointer.
-				 */
-				AE_SP24F_L_C(d, (ae_p24f *)fir->fir_wp, n_sz);
+				AE_L16_XP(d, (ae_int16 *)x_rptr,
+					  sizeof(ae_int16));
+				q = AE_CVT32X2F16_32(d);
+				AE_S32_L_XC(q, (ae_int32 *)fir->fir_wp, n_sz);
 			}
 
 			/* Check for wrap */
@@ -521,17 +523,17 @@ void src_polyphase_stage_cir_s16(struct src_stage_prm *s)
 
 		/* Do filter */
 		cp = cfg->coefs; /* Reset to 1st coefficient */
-		rp = (ae_q32s *)fir->fir_wp;
+		rp = (ae_f32 *)fir->fir_wp;
 
 		/* Do circular modification to pointer rp by amount of
-		 * rewind to to data start. Loaded value d is discarded.
+		 * rewind to to data start. Loaded value q is discarded.
 		 */
-		AE_LP24F_C(d, (ae_p24f *)rp, rewind_sz);
+		AE_L32_XC(q, rp, rewind_sz);
 
 		/* Reset FIR output write pointer and compute all polyphase
 		 * sub-filters.
 		 */
-		wp = (ae_q32s *)fir->out_rp;
+		wp = (ae_f32 *)fir->out_rp;
 		for (i = 0; i < cfg->num_of_subfilters; i++) {
 			fir_filter(rp, cp, wp, taps_div_4, cfg->shift, nch);
 			wp += nch_x_odm;
@@ -539,10 +541,10 @@ void src_polyphase_stage_cir_s16(struct src_stage_prm *s)
 			src_inc_wrap((int32_t **)&wp, out_delay_end, out_size);
 
 			/* Circular advance pointer rp by number of
-			 * channels x input delay multiplier. Loaded value d
+			 * channels x input delay multiplier. Loaded value q
 			 * is discarded.
 			 */
-			AE_LP24F_C(d, (ae_p24f *)rp, nch_x_idm_sz);
+			AE_L32_XC(q, rp, nch_x_idm_sz);
 		}
 
 		/* Output */
@@ -556,18 +558,19 @@ void src_polyphase_stage_cir_s16(struct src_stage_prm *s)
 			n_min = (m < n_wrap_buf) ? m : n_wrap_buf;
 			m -= n_min;
 			for (i = 0; i < n_min; i++) {
-				/* Circular load for 32 bit sample, get 24
-				 * high bits, circular advance pointer.
+				/* Circular load for 32 bit sample,
+				 * advance read pointer.
 				 */
-				AE_LP24F_C(d, (ae_p24f *)fir->out_rp, sz);
+				AE_L32_XC(q, (ae_int32 *)fir->out_rp, sz);
 
-				/* Round Q1.12 value and store as Q1.15 and
-				 * advance pointer.
+				/* Store Q1.31 value as Q1.15 and
+				 * advance write pointer.
 				 */
-				d = AE_ROUNDSP16SYM(d);
-				AE_SP16F_L_I(d, (ae_p16s *)y_wptr, 0);
-				y_wptr++;
+				d = AE_ROUND16X4F32SSYM(q, q);
+				AE_S16_0_XP(d, (ae_int16 *)y_wptr,
+					    sizeof(ae_int16));
 			}
+
 			/* Check wrap */
 			src_inc_wrap_s16(&y_wptr, y_end_addr, s->y_size);
 		}
