+++ src/audio/volume/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_local_sources(sof volume_generic.c volume_hifi3.c volume.c)
