+++ src/include/sof/audio/src/src_config.h
