+++ src/include/sof/audio/volume.h
